+++ components/Footer.tsx
@@ -1,0 +1,11 @@
+const Footer = () => {
+  return (
+    <footer className="mt-auto">
+      <p className="text-center">
+        © 2022 Alvito Raidhy Trinanda. MIT Licensed.
+      </p>
+    </footer>
+  );
+};
+
+export default Footer;