+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 <!-- markdownlint-disable MD024-->
+
 # **Change Log** 📜📝
 
 All notable changes to the "**noturist**" web will be documented in this file.
@@ -7,8 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ---
 
+## [**0.1.0**] - 2022-01-25
+
+### Added
+
+- Register and login functionality
+- Create, read, update, and delete notes feature
+
 ## [**0.0.1**] - 2022-01-21
 
 ### Added
 
-* The basic project structure from **[josee9988/project-template](https://github.com/Josee9988/project-template)**.
+- The basic project structure from **[josee9988/project-template](https://github.com/Josee9988/project-template)**.